fix php8.1 gensalt_blowfish E_DEPRECATED
fix > Deprecated: Implicit conversion from float 48.8 to int loses precision

trac: https://core.trac.wordpress.org/ticket/56340

+++ src/wp-includes/class-phpass.php
@@ -174,7 +174,7 @@ class PasswordHash {
 
 		$output = '$2a$';
 		$output .= chr(ord('0') + $this->iteration_count_log2 / 10);
-		$output .= chr(ord('0') + $this->iteration_count_log2 % 10);
+		$output .= chr((int)(ord('0') + $this->iteration_count_log2 % 10));
 		$output .= '$';
 
 		$i = 0;
